Move code to handle snapshot filenames
